@@ -1,15 +1,28 @@
 module ih_server
 
-go 1.16
+go 1.20
 
 require (
-	github.com/go-sql-driver/mysql v1.4.1-0.20190127011111-1b9eda25029c
-	github.com/golang/protobuf v1.3.1
+	github.com/go-sql-driver/mysql v1.7.0
+	github.com/golang/protobuf v1.5.3
 	github.com/golang/snappy v0.0.0-20180518054509-2e65f85255db
-	github.com/gomodule/redigo v2.0.1-0.20180627144507-2cd21d9966bf+incompatible
+	github.com/gomodule/redigo v1.8.9
+	github.com/hashicorp/golang-lru v0.5.4
+	github.com/huoshan017/mysql-go v0.0.0-20230415101517-a60dd29245de
+	github.com/huoshan017/ponu v0.0.0-20230418061437-64be33c5807e
 	github.com/orcaman/concurrent-map v0.0.0-20210501183033-44dafcb38ecc
 	github.com/satori/go.uuid v1.2.1-0.20181028125025-b2ce2384e17b
 	github.com/tealeg/xlsx v0.0.0-20181024002044-dbf71b6a931e
-	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.30.0
+)
+
+require (
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/onsi/ginkgo v1.16.5 // indirect
+	github.com/onsi/gomega v1.27.6 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/redis/go-redis/v9 v9.0.3 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
