@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"ih_server/libs/log"
 	"ih_server/src/server_config"
-	"io/ioutil"
+	"os"
 )
 
 type GlobalConfig struct {
@@ -31,7 +31,7 @@ type GlobalConfig struct {
 var global_config GlobalConfig
 
 func global_config_load() bool {
-	data, err := ioutil.ReadFile(server_config.GetGameDataPathFile("global.json"))
+	data, err := os.ReadFile(server_config.GetGameDataPathFile("global.json"))
 	if nil != err {
 		log.Error("global_config_load failed to readfile err(%s)!", err.Error())
 		return false
