@@ -34,123 +34,123 @@ func NewCommonRankingList(root_node SkiplistNode, max_rank int32) *CommonRanking
 	return ranking_list
 }
 
-func (this *CommonRankingList) GetByRank(rank int32) SkiplistNode {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetByRank(rank int32) SkiplistNode {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	item := this.ranking_items.GetByRank(rank)
+	item := list.ranking_items.GetByRank(rank)
 	if item == nil {
 		return nil
 	}
-	new_item := this.items_pool.Get().(SkiplistNode)
+	new_item := list.items_pool.Get().(SkiplistNode)
 	new_item.Assign(item)
 	return new_item
 }
 
-func (this *CommonRankingList) GetByKey(key interface{}) SkiplistNode {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetByKey(key interface{}) SkiplistNode {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	item, o := this.key2item[key]
+	item, o := list.key2item[key]
 	if !o || item == nil {
 		return nil
 	}
-	new_item := this.items_pool.Get().(SkiplistNode)
+	new_item := list.items_pool.Get().(SkiplistNode)
 	new_item.Assign(item)
 	return new_item
 }
 
-func (this *CommonRankingList) SetValueByKey(key interface{}, value interface{}) {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) SetValueByKey(key interface{}, value interface{}) {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	item, o := this.key2item[key]
+	item, o := list.key2item[key]
 	if !o || item == nil {
 		return
 	}
 	item.SetValue(value)
 }
 
-func (this *CommonRankingList) insert(key interface{}, item SkiplistNode, is_lock bool) bool {
+func (list *CommonRankingList) insert(key interface{}, item SkiplistNode, is_lock bool) bool {
 	if is_lock {
-		this.locker.Lock()
-		defer this.locker.Unlock()
+		list.locker.Lock()
+		defer list.locker.Unlock()
 	}
-	this.ranking_items.Insert(item)
-	this.key2item[key] = item
+	list.ranking_items.Insert(item)
+	list.key2item[key] = item
 	return true
 }
 
-func (this *CommonRankingList) Insert(key interface{}, item SkiplistNode) bool {
-	return this.insert(key, item, true)
+func (list *CommonRankingList) Insert(key interface{}, item SkiplistNode) bool {
+	return list.insert(key, item, true)
 }
 
-func (this *CommonRankingList) delete(key interface{}, is_lock bool) bool {
+func (list *CommonRankingList) delete(key interface{}, is_lock bool) bool {
 	if is_lock {
-		this.locker.Lock()
-		defer this.locker.Unlock()
+		list.locker.Lock()
+		defer list.locker.Unlock()
 	}
 
-	item, o := this.key2item[key]
+	item, o := list.key2item[key]
 	if !o {
 		log.Error("CommonRankingList key[%v] not found", key)
 		return false
 	}
-	if !this.ranking_items.Delete(item) {
+	if !list.ranking_items.Delete(item) {
 		log.Error("CommonRankingList delete key[%v] value[%v] in ranking list failed", key, item.GetValue())
 		return false
 	}
 	if is_lock {
-		this.items_pool.Put(item)
+		list.items_pool.Put(item)
 	}
-	//delete(this.key2item, key)
+	//delete(list.key2item, key)
 	return true
 }
 
-func (this *CommonRankingList) Delete(key interface{}) bool {
-	return this.delete(key, true)
+func (list *CommonRankingList) Delete(key interface{}) bool {
+	return list.delete(key, true)
 }
 
-func (this *CommonRankingList) Update(item SkiplistNode) bool {
-	this.locker.Lock()
-	defer this.locker.Unlock()
+func (list *CommonRankingList) Update(item SkiplistNode) bool {
+	list.locker.Lock()
+	defer list.locker.Unlock()
 
-	old_item, o := this.key2item[item.GetKey()]
+	old_item, o := list.key2item[item.GetKey()]
 	if o {
-		if !this.delete(item.GetKey(), false) {
+		if !list.delete(item.GetKey(), false) {
 			log.Error("Update key[%v] for Ranking List failed", item)
 			return false
 		}
 		old_item.Assign(item)
-		return this.insert(item.GetKey(), old_item, false)
+		return list.insert(item.GetKey(), old_item, false)
 	} else {
-		new_item := this.items_pool.Get().(SkiplistNode)
+		new_item := list.items_pool.Get().(SkiplistNode)
 		new_item.Assign(item)
-		return this.insert(item.GetKey(), new_item, false)
+		return list.insert(item.GetKey(), new_item, false)
 	}
 }
 
-func (this *CommonRankingList) GetRangeNodes(rank_start, rank_num int32, nodes []interface{}) (num int32) {
-	if rank_start <= int32(0) || rank_start > this.max_rank {
+func (list *CommonRankingList) GetRangeNodes(rank_start, rank_num int32, nodes []interface{}) (num int32) {
+	if rank_start <= int32(0) || rank_start > list.max_rank {
 		log.Warn("Ranking list rank_start[%v] invalid", rank_start)
 		return
 	}
 
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	if int(rank_start) > len(this.key2item) {
-		log.Warn("Ranking List rank range[1,%v], rank_start[%v] over rank list", len(this.key2item), rank_start)
+	if int(rank_start) > len(list.key2item) {
+		log.Warn("Ranking List rank range[1,%v], rank_start[%v] over rank list", len(list.key2item), rank_start)
 		return
 	}
 
-	real_num := int32(len(this.key2item)) - rank_start + 1
+	real_num := int32(len(list.key2item)) - rank_start + 1
 	if real_num < rank_num {
 		rank_num = real_num
 	}
 
 	items := make([]SkiplistNode, rank_num)
-	b := this.ranking_items.GetByRankRange(rank_start, rank_num, items)
+	b := list.ranking_items.GetByRankRange(rank_start, rank_num, items)
 	if !b {
 		log.Warn("Ranking List rank range[%v,%v] is empty", rank_start, rank_num)
 		return
@@ -169,35 +169,35 @@ func (this *CommonRankingList) GetRangeNodes(rank_start, rank_num int32, nodes [
 	return
 }
 
-func (this *CommonRankingList) GetRank(key interface{}) int32 {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetRank(key interface{}) int32 {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	item, o := this.key2item[key]
+	item, o := list.key2item[key]
 	if !o {
 		return 0
 	}
-	return this.ranking_items.GetRank(item)
+	return list.ranking_items.GetRank(item)
 }
 
-func (this *CommonRankingList) GetRankAndValue(key interface{}) (rank int32, value interface{}) {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetRankAndValue(key interface{}) (rank int32, value interface{}) {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	item, o := this.key2item[key]
+	item, o := list.key2item[key]
 	if !o {
 		return 0, nil
 	}
 
-	return this.ranking_items.GetRank(item), item.GetValue()
+	return list.ranking_items.GetRank(item), item.GetValue()
 }
 
-func (this *CommonRankingList) GetRankRange(start, num int32) (int32, int32) {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetRankRange(start, num int32) (int32, int32) {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	l := int32(len(this.key2item))
-	if this.key2item == nil || l == 0 {
+	l := int32(len(list.key2item))
+	if list.key2item == nil || l == 0 {
 		return 0, 0
 	}
 
@@ -211,12 +211,12 @@ func (this *CommonRankingList) GetRankRange(start, num int32) (int32, int32) {
 	return start, num
 }
 
-func (this *CommonRankingList) GetLastRankRange(num int32) (int32, int32) {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetLastRankRange(num int32) (int32, int32) {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	l := int32(len(this.key2item))
-	if this.key2item == nil || l == 0 {
+	l := int32(len(list.key2item))
+	if list.key2item == nil || l == 0 {
 		return 0, 0
 	}
 
@@ -227,15 +227,15 @@ func (this *CommonRankingList) GetLastRankRange(num int32) (int32, int32) {
 	return l - num + 1, num
 }
 
-func (this *CommonRankingList) GetLastRank() int32 {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
+func (list *CommonRankingList) GetLastRank() int32 {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
 
-	return int32(len(this.key2item))
+	return int32(len(list.key2item))
 }
 
-func (this *CommonRankingList) GetLength() int32 {
-	this.locker.RLock()
-	defer this.locker.RUnlock()
-	return this.ranking_items.GetLength()
+func (list *CommonRankingList) GetLength() int32 {
+	list.locker.RLock()
+	defer list.locker.RUnlock()
+	return list.ranking_items.GetLength()
 }
