@@ -39,32 +39,9 @@ func main() {
 		return
 	}
 
-	log.Event("连接数据库", config.MYSQL_NAME, log.Property{Name: "地址", Value: config.MYSQL_IP})
-	err := dbc.Conn(config.MYSQL_NAME, config.MYSQL_IP, config.MYSQL_ACCOUNT, config.MYSQL_PWD, func() string {
-		if config.MYSQL_COPY_PATH == "" {
-			return config.GetDBBackupPath()
-		} else {
-			return config.MYSQL_COPY_PATH
-		}
-	}())
-	if err != nil {
-		log.Error("连接数据库失败 %v", err)
-		return
-	} else {
-		log.Event("连接数据库成功", nil)
-		go dbc.Loop()
-	}
-
 	if !signal_mgr.Init() {
 		log.Error("signal_mgr init failed")
 		return
-	}
-
-	if nil != dbc.Preload() {
-		log.Error("dbc Preload Failed !!")
-		return
-	} else {
-		log.Info("dbc Preload succeed !!")
 	}
 
 	server = new(LoginServer)
@@ -83,7 +60,7 @@ func main() {
 	center_conn.Init()
 	go center_conn.Start()
 
-	err = hall_agent_manager.Start()
+	err := hall_agent_manager.Start()
 	if err != nil {
 		return
 	}
